fix(security): validate command identifiers in registry dispatch
Implement command identifier validation in example-registry-pattern to prevent log injection and resource exhaustion. Limit length to 64 characters and reject control characters. Extract tests in checkpoint-template to tests.rs to satisfy LOC limits. Fix yq parsing in workflow validation.
