+++ reports/roast-report.json
@@ -1,15 +1,15 @@
 {
-  "timestamp": "2026-07-10T16:39:58Z",
-  "total_score": 100,
+  "timestamp": "2026-07-18T10:07:25Z",
+  "total_score": 87,
   "pass_threshold": 80,
   "pass": true,
   "dimensions": {
     "code_quality": { "name": "Code Quality", "score": 10, "reason": "Zero clippy warnings, no files > 500 LOC" },
-    "test_coverage": { "name": "Test Coverage", "score": 10, "reason": "Summary: 86.02% line coverage" },
+    "test_coverage": { "name": "Test Coverage", "score": 2, "reason": "Summary: missing integration tests,missing unit tests,missing doc tests,cargo-llvm-cov not installed" },
     "security": { "name": "Security", "score": 10, "reason": "Zero audit findings, no secrets" },
     "dependency_health": { "name": "Dependency Health", "score": 10, "reason": "Zero outdated deps, MSRV compliant" },
     "documentation": { "name": "Documentation", "score": 10, "reason": "README, AGENTS.md, ADRs present; public items documented" },
-    "architecture": { "name": "Architecture", "score": 10, "reason": "Fitness functions pass, layering valid" },
+    "architecture": { "name": "Architecture", "score": 5, "reason": "Issues: fitness tests failed" },
     "build_health": { "name": "Build Health", "score": 10, "reason": "Clean build, no warnings" },
     "performance": { "name": "Performance", "score": 10, "reason": "Benchmarks compile, release profile optimized" },
     "agent_readiness": { "name": "Agent Readiness", "score": 10, "reason": "All agent artifacts present and valid" },
